Update TOPICS.md
If we look precisely at make_point_cloud() function, the order of parameters is the following : relative_velocity, SnR and then ground_velocity. 
Here, in the table, the order is relative_velocity, ground_velocity, SnR.

It confused me this afternoon, as I use this stable as Ground Truth of the fields position, so I propose to correct the order in the table.